fix: 한 id 에 걸린 버스트 전부의 마감을 다시 센다 · BLE 회수 시계를 부호 있게 · 없앤 권한 축을 아직 있다고 말하던 일곱 곳
2026-09-22 전수 점검의 D, E, 그리고 7번(보드 쪽).

1. fsd_burst - 발화한 슬롯만 마감을 다시 세고 있었다 (점검 D)
   마감이 답하는 질문은 "이 id 가 조용해졌나" 이고(FSD_BURST_MAX_WAIT_MS 의
   주석이 그렇게 적는다), 그것은 어느 슬롯을 골랐는지와 무관하다. 그런데
   고른 슬롯만 다시 셌다.

   맵등과 미러는 같은 0x273 을 타고 한 번 누름이 둘을 함께 무장할 수 있다.
   0x273 은 500 ms 주기이고 마감은 1000 ms 라, 둘째 슬롯은 두 번째 도착까지
   살아남아야 했고 지터에 여유가 없었다. 지면 "대기 중이던 명령을 버렸다"
   가 찍히는데 버스는 한 번도 조용한 적이 없다 - 열두 번째 패턴 그대로다.

   RED 를 먼저 봤다: 고침을 빼면 새 시험 4줄이 빨개진다.

2. ble_server - 회수 유예를 부호 있는 차로 잰다 (점검 E)
   g_link_down_ms 는 onDisconnect 가 BLE 호스트 태스크에서 찍고, now_ms 는
   loop() 가 한참 앞에서 잰 값이다. 끊김이 그 사이에 떨어지면 stamp 가 now
   보다 앞서고, 부호 없이 빼면 30초 유예가 0초가 된다. 2026-09-10 에 OTA 가
   정확히 이 산수에 물렸는데(fsd_ota_stalled) 여기만 안 따라왔다.

   .firmware 는 호스트 시험이 안 닿는다. 그래서 이것은 시험이 아니라 근거로
   남긴다 - 고친 자리의 형제이고, 틀리면 덜 쏘는 쪽으로 틀린다.

3. 없앤 권한 축을 아직 "있다" 고 말하던 일곱 곳 (점검 7)
   fsd_body_allows() 와 armable_at_runtime 은 2026-09-10 에 없앴는데, 일곱
   곳이 현재형으로 그것을 가리키고 있었다. 가장 나쁜 둘:

   - fsd_pipeline.h 의 "WHAT IS STILL ASKED HERE" 표가 없는 게이트 여덟을
     약속하고 있었다.
   - fsd_body_t2.h / .c 는 문 열기를 두고 "권한 쪽은 이미 닫혀 있다" 고
     적었는데, 그 플래그도 없고 문 방출기는 2026-09-05 부터 있다.

   문장을 지우지 않고 "무엇이 틀렸었나" 를 같이 남겼다. 그리고 T2 에는
   지금 실제로 무엇이 막고 있는지(부르는 곳이 없다)와, 그것을 이으면
   무엇이 남는지를 적었다 - 차주가 정할 일이다.

   나머지 아홉 자리는 이미 "없앴다" 로 제대로 적혀 있어 안 건드렸다.

확인: 호스트 시험 18묶음 5,560건(+10) 실패 0, 보드 8종 SUCCESS, 경고 0.

+++ esp32/.firmware/ble_server.cpp
@@ -918,7 +918,20 @@ static void ble_revoke_active_if_stale(uint32_t now_ms) {
     // g_owner_present, not g_connected: a stranger holding the link open is not
     // a reason to leave the owner's transmit permission standing. See the flag.
     if (g_owner_present || g_link_down_ms == 0) return;
-    if ((uint32_t)(now_ms - g_link_down_ms) < BLE_ACTIVE_GRACE_MS) return;
+    /* 🔴 **부호 있는 차로 잰다.** g_link_down_ms 는 onDisconnect 가 **BLE
+     * 호스트 태스크**에서 자기 millis() 로 찍고(아래 ServerCB), now_ms 는
+     * loop() 가 한참 앞에서 재 둔 값이다(main.cpp). 끊김이 그 사이에 떨어지면
+     * stamp 가 now 보다 **앞선다** — 부호 없이 빼면 그 차가 약 4,290,000,000 ms
+     * 로 읽혀 30초 유예가 **0초**가 되고, 폰이 잠깐 끊긴 것만으로 Active 가
+     * 즉시 회수된다.
+     *
+     * 🔴 이 저장소는 2026-09-10 에 OTA 에서 **정확히 이 산수**에 반나절을 썼다
+     * (fsd_ota_stalled() 이 그때 부호 있는 차로 바뀐 자리다). 여기만 안 따라
+     * 왔다 — 고친 곳의 형제를 안 고치는 일곱 번째 패턴이다.
+     *
+     * 🟢 앞선 stamp 는 «방금» 으로 읽힌다: 음수 차는 이 비교에서 참이 되어
+     * 기다리는 쪽으로 떨어진다. */
+    if ((int32_t)(now_ms - g_link_down_ms) < (int32_t)BLE_ACTIVE_GRACE_MS) return;
 
     // 🔴 The bookkeeping is cleared AFTER the switch succeeds, not before.
     //
